Channels existence fetch hook
