+++ apps/products-feed/src/pages/categories.tsx
@@ -4,16 +4,10 @@ import { trpcClient } from "../modules/trpc/trpc-client";
 import { useRouter } from "next/router";
 import { ConfigurationPageBaseLayout } from "../modules/ui/configuration-page-base-layout";
 import { CategoryMapping } from "../modules/category-mapping/ui/category-mapping";
+import { useChannelsExistenceChecking } from "../modules/channels/use-channels-existence-checking";
 
 const ConfigurationPage: NextPage = () => {
-  const channels = trpcClient.channels.fetch.useQuery();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (channels.isSuccess && channels.data.length === 0) {
-      router.push("/not-ready");
-    }
-  }, [channels.data, channels.isSuccess]);
+  useChannelsExistenceChecking();
 
   return (
     <ConfigurationPageBaseLayout>
